move enums to seprate py, create defaults
this change allows modyfing default in single place
this allows for easier debug of new features
and adjusting defaults
